+++ src/loan-service/src/main/java/com/asc/loanservice/domain/loan/application/Loan.java
@@ -1,7 +1,9 @@
 package com.asc.loanservice.domain.loan.application;
 
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 import javax.persistence.Embeddable;
 import java.math.BigDecimal;
@@ -10,10 +12,12 @@ import java.time.LocalDate;
 
 @Getter
 @Builder
+@AllArgsConstructor
+@NoArgsConstructor
 @Embeddable
 class Loan {
 
-    private final BigDecimal loanAmount;
-    private final Integer numberOfInstallments;
-    private final LocalDate firstInstallmentDate;
+    private BigDecimal loanAmount;
+    private Integer numberOfInstallments;
+    private LocalDate firstInstallmentDate;
 }
