+++ src/loan-service/src/main/java/com/asc/loanservice/domain/loan/application/LoanApplicationRepository.java
@@ -1,6 +1,5 @@
 package com.asc.loanservice.domain.loan.application;
 
-import com.asc.loanservice.infrastructure.repository.model.LoanApplication;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
